requirements.txt
Pillow>=10.0.0
tqdm>=4.66.0
numpy>=1.24.0
rfdetr>=1.6.0
